Create readme.md
Open Seadragon Map Polygon Annotation
This is an example of polygon annotations as overlays on an openseadragon image viewer. Polygons are defined as an array of x and y coordinates. For each annotation, the user can define a description that is shown aside of the polygon annotation (active/deactivate description by clicking on the polygons)